requirements.txt
trl==0.29.0
transformers>=4.40.0
peft>=0.10.0
vllm>=0.6.0
fastapi
uvicorn[standard]
httpx
anthropic
python-dotenv
torch>=2.2.0
pydantic>=2.0
